Moved declaritive base to db package.
