+++ tds/db/__init__.py
@@ -2,8 +2,12 @@
 tds.db - DB handling
 """
 
+from sqlalchemy.ext.declarative import declarative_base
+
 from tds.db.elasticsearch import es_client
 from tds.db.graph.neo4j import request_engine as request_graph_db
 from tds.db.helpers import drop_content, entry_exists, init_dev_content, list_by_id
 from tds.db.relational import engine as rdb
 from tds.db.relational import request_engine as request_rdb
+
+Base = declarative_base()
