+++ tds/modules/person/model.py
@@ -7,7 +7,7 @@ import sqlalchemy as sa
 from pydantic import BaseModel
 from sqlalchemy.orm import relationship
 
-from tds.autogen.orm import Base
+from tds.db import Base
 from tds.db.enums import ResourceType, Role
