+++ tds/modules/provenance/model.py
@@ -9,7 +9,7 @@ from pydantic import Field
 from pydantic.main import BaseModel
 from sqlalchemy import func
 
-from tds.autogen.orm import Base
+from tds.db import Base
 from tds.db.enums import ProvenanceType, RelationType
